refactor: PaymentClient 추상화
